phenomenon is singular, phenomena is plural
As with other plurals of Latin or Greek origin, like media and criteria, there is a tendency to use the plural phenomena as a singular ( This phenomena will not be seen again), but such use occurs infrequently in edited writing. The plural form phenomenas, though occasionally seen, has even less currency.